Refactor players to model folder
