+++ Controllers/HomeController.cs
@@ -54,14 +54,7 @@ public class HomeController : Controller
     }
 
 
-    public class Player
-    {
-        public int PlayerID { get; set; }
-        public string? PlayerFirstName { get; set; }
-        public int Age { get; set; }
-        public object? ViewBag { get; set; }
-        public string? TotalPlayers { get; set; }
-    }
+
