+++ Models/Player.cs
@@ -1,0 +1,12 @@
+
+namespace dotnetNewProject.Models;
+
+
+public class Player
+{
+    public int PlayerID { get; set; }
+    public string? PlayerFirstName { get; set; }
+    public int Age { get; set; }
+    public object? ViewBag { get; set; }
+    public string? TotalPlayers { get; set; }
+}
