+++ Views/Home/Leaderboard.cshtml
@@ -1,4 +1,6 @@
-﻿@using static dotnetNewProject.Controllers.HomeController
+﻿@using dotnetNewProject.Models
+
+
 
 
 <ul>
